+++ src/Pagebuilder/Exceptions/MissingRequiredFieldsExeption.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Flobbos\Pagebuilder\Exceptions;
-
-use Exception;
-
-class MissingRequiredFieldsException extends Exception{
-    
-}